+++ src/main/kotlin/ch/skyfy/tinyeconomyrenewed/server/config/EarnMoneyByKillingPlayersConfig.kt
@@ -1,12 +1,12 @@
 package ch.skyfy.tinyeconomyrenewed.server.config
 
-import ch.skyfy.json5configlib.Defaultable
-import ch.skyfy.json5configlib.Validatable
+import ch.skyfy.jsonconfiglib.Defaultable
+import ch.skyfy.jsonconfiglib.Validatable
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class EarnMoneyByKillingPlayersConfig(val amount: Float, val shouldKilledPlayerLostMoney: Boolean) : Validatable
+data class EarnMoneyByKillingPlayersConfig(val amount: Double, val shouldKilledPlayerLostMoney: Boolean) : Validatable
 
 class DefaultEarnMoneyByKillingPlayersConfig: Defaultable<EarnMoneyByKillingPlayersConfig> {
-    override fun getDefault() = EarnMoneyByKillingPlayersConfig(10f, true)
+    override fun getDefault() = EarnMoneyByKillingPlayersConfig(10.0, true)
 }
