+++ src/main/kotlin/ch/skyfy/tinyeconomyrenewed/server/config/EarnMoneyLoginConfig.kt
@@ -1,12 +1,12 @@
 package ch.skyfy.tinyeconomyrenewed.server.config
 
-import ch.skyfy.json5configlib.Defaultable
-import ch.skyfy.json5configlib.Validatable
+import ch.skyfy.jsonconfiglib.Defaultable
+import ch.skyfy.jsonconfiglib.Validatable
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class EarnMoneyLoginConfig(val amount: Float) : Validatable
+data class EarnMoneyLoginConfig(val amount: Double) : Validatable
 
 class DefaultEarnMoneyLoginConfig : Defaultable<EarnMoneyLoginConfig> {
-    override fun getDefault() = EarnMoneyLoginConfig(10f)
+    override fun getDefault() = EarnMoneyLoginConfig(10.0)
 }
