+++ src/main/kotlin/ch/skyfy/tinyeconomyrenewed/server/config/LossMoneyDyingConfig.kt
@@ -1,12 +1,12 @@
 package ch.skyfy.tinyeconomyrenewed.server.config
 
-import ch.skyfy.json5configlib.Defaultable
-import ch.skyfy.json5configlib.Validatable
+import ch.skyfy.jsonconfiglib.Defaultable
+import ch.skyfy.jsonconfiglib.Validatable
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class LossMoneyDyingConfig(val amount: Float) : Validatable
+data class LossMoneyDyingConfig(val amount: Double) : Validatable
 
 class DefaultLossMoneyDyingConfig : Defaultable<LossMoneyDyingConfig> {
-    override fun getDefault() = LossMoneyDyingConfig(50f)
+    override fun getDefault() = LossMoneyDyingConfig(50.0)
 }
