+++ src/main/kotlin/ch/skyfy/tinyeconomyrenewed/server/config/VillagerTradeCostsMoneyConfig.kt
@@ -1,17 +1,17 @@
 package ch.skyfy.tinyeconomyrenewed.server.config
 
-import ch.skyfy.json5configlib.Defaultable
-import ch.skyfy.json5configlib.Validatable
+import ch.skyfy.jsonconfiglib.Defaultable
+import ch.skyfy.jsonconfiglib.Validatable
 
 @kotlinx.serialization.Serializable
 @JvmRecord
 data class VillagerTradeCostsMoneyConfig(
     @JvmField
     val enabled: Boolean,
-    val price: Float,
+    val price: Double,
     val amount: Int
 ): Validatable
 
 class DefaultVillagerTradeCostsMoneyConfig : Defaultable<VillagerTradeCostsMoneyConfig>{
-    override fun getDefault() = VillagerTradeCostsMoneyConfig(false, 0.5f, 2)
+    override fun getDefault() = VillagerTradeCostsMoneyConfig(false, 0.5, 2)
 }
