+++ src/main/resources/assets/tinyeconomyrenewed/sql/init.sql
@@ -34,7 +34,7 @@ create table if not exists player
     id int auto_increment primary key,
     uuid  varchar(255) not null,
     name  varchar(255) not null,
-    money float default 0 not null,
+    money double default 0 not null,
     constraint player_uuid_key_uindex unique (uuid)
 );
 
@@ -42,7 +42,7 @@ create table if not exists mined_block_reward
 (
     id int auto_increment primary key,
     block_id int not null,
-    amount float null,
+    amount double null,
     constraint mined_block_reward_block__fk foreign key (block_id) references block (id),
     constraint mined_block_reward_block_id_uindex unique (block_id)
 );
@@ -51,7 +51,7 @@ create table if not exists entity_killed_reward
 (
     id int auto_increment primary key,
     entity_id int not null,
-    amount float null,
+    amount double null,
     constraint entity_killed_reward_entity__fk foreign key (entity_id) references entity (id),
     constraint entity_killed_reward_entity_id_uindex unique (entity_id)
 );
@@ -60,7 +60,7 @@ create table if not exists advancement_reward
 (
     id int auto_increment primary key,
     advancement_id int not null,
-    amount float null,
+    amount double null,
     constraint advancement_reward_advancement__fk foreign key (advancement_id) references advancement (id),
     constraint advancement_reward_advancement_id_uindex unique (advancement_id)
 );
